+++ pytest_factoryboy/__init__.py
@@ -1,10 +1,7 @@
 """pytest-factoryboy public API."""
 from .fixture import register, LazyFixture
 
-__version__ = "2.1.0"
+__version__ = "2.2.0"
 
 
-__all__ = [
-    register.__name__,
-    LazyFixture.__name__,
-]
+__all__ = ("register", "LazyFixture")