+++ tests/test_postgen_dependencies.py
@@ -1,34 +1,54 @@
-"""Test post-generation dependecies."""
+"""Test post-generation dependencies."""
+from __future__ import annotations
+
+from dataclasses import dataclass
 
 import factory
 import pytest
 
 from factory.declarations import NotProvided
 from pytest_factoryboy import register
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from typing import Any
+    from pytest_factoryboy.plugin import Request
 
 
+@dataclass
 class Foo:
-    def __init__(self, value, expected):
-        self.value = value
-        self.expected = expected
-        self.secret = ""
-        self.number = 4
+    value: int
+    expected: int
+    secret: str = ""
+    number: int = 4
 
-    def set_secret(self, new_secret):
+    def set_secret(self, new_secret: str) -> None:
         self.secret = new_secret
 
-    def set_number(self, new_number=987):
+    def set_number(self, new_number: int = 987) -> None:
         self.number = new_number
 
 
+@dataclass
 class Bar:
-    def __init__(self, foo):
-        self.foo = foo
+    foo: Foo
+
+
+@dataclass
+class Baz:
+    foo: Foo
+
+
+@register
+class BazFactory(factory.Factory):
+    class Meta:
+        model = Baz
+
+    foo = None
 
 
 @register
 class FooFactory(factory.Factory):
-
     """Foo factory."""
 
     class Meta:
@@ -37,12 +57,16 @@ class FooFactory(factory.Factory):
     value = 0
     #: Value that is expected at the constructor
     expected = 0
+
     secret = factory.PostGenerationMethodCall("set_secret", "super secret")
     number = factory.PostGenerationMethodCall("set_number")
 
     @factory.post_generation
-    def set1(foo, create, value, **kwargs):
+    def set1(foo: Foo, create: bool, value: Any, **kwargs: Any) -> str:
         foo.value = 1
+        return "set to 1"
+
+    baz = factory.RelatedFactory(BazFactory, factory_related_name="foo")
 
     @factory.post_generation
     def set2(foo, create, value, **kwargs):
@@ -50,19 +74,18 @@ class FooFactory(factory.Factory):
             foo.value = value
 
     @classmethod
-    def _after_postgeneration(cls, obj, create, results=None):
+    def _after_postgeneration(cls, obj: Foo, create: bool, results: dict[str, Any] | None = None) -> None:
         obj._postgeneration_results = results
         obj._create = create
 
 
 class BarFactory(factory.Factory):
-
     """Bar factory."""
 
     foo = factory.SubFactory(FooFactory)
 
     @classmethod
-    def _create(cls, model_class, foo):
+    def _create(cls, model_class: type[Bar], foo: Foo) -> Bar:
         assert foo.value == foo.expected
         bar = super()._create(model_class, foo=foo)
         foo.bar = bar
@@ -72,7 +95,7 @@ class BarFactory(factory.Factory):
         model = Bar
 
 
-def test_postgen_invoked(foo):
+def test_postgen_invoked(foo: Foo):
     """Test that post-generation hooks are done and the value is 2."""
     assert foo.value == 1
 
@@ -82,12 +105,12 @@ register(BarFactory)
 
 @pytest.mark.parametrize("foo__value", [3])
 @pytest.mark.parametrize("foo__expected", [1])
-def test_depends_on(bar):
+def test_depends_on(bar: Bar):
     """Test that post-generation hooks are done and the value is 1."""
     assert bar.foo.value == 1
 
 
-def test_getfixturevalue(request, factoryboy_request):
+def test_getfixturevalue(request, factoryboy_request: Request):
     """Test post-generation declarations via the getfixturevalue."""
     foo = request.getfixturevalue("foo")
     assert not factoryboy_request.deferred
@@ -114,14 +137,15 @@ def test_postgeneration_getfixturevalue(request, factoryboy_request):
     assert set2 is None
 
 
-def test_after_postgeneration(foo):
+def test_after_postgeneration(foo: Foo):
     """Test _after_postgeneration is called."""
-    assert foo._postgeneration_results == {"set1": None, "set2": None, "secret": None, "number": None}
     assert foo._create is True
+    assert foo._postgeneration_results.pop("baz")
+    assert foo._postgeneration_results == {"set1": "set to 1", "set2": None, "secret": None, "number": None}
 
 
 @pytest.mark.parametrize("foo__set2", [123])
-def test_postgeneration_fixture(foo):
+def test_postgeneration_fixture(foo: Foo):
     """Test fixture for ``PostGeneration`` declaration."""
     assert foo.value == 123
 
@@ -132,7 +156,7 @@ def test_postgeneration_fixture(foo):
         ("test secret", 456),
     ],
 )
-def test_postgenerationmethodcall_fixture(foo):
+def test_postgenerationmethodcall_fixture(foo: Foo):
     """Test fixture for ``PostGenerationMethodCall`` declaration."""
     assert foo.secret == "test secret"
     assert foo.number == 456
@@ -148,14 +172,14 @@ class OrderedFactory(factory.Factory):
         model = Ordered
 
     @factory.post_generation
-    def zzz(obj, create, val, **kwargs):
+    def zzz(obj: Ordered, create: bool, val: Any, **kwargs: Any) -> None:
         obj.value = "zzz"
 
     @factory.post_generation
-    def aaa(obj, create, val, **kwargs):
+    def aaa(obj: Ordered, create: bool, val: Any, **kwargs: Any) -> None:
         obj.value = "aaa"
 
 
-def test_ordered(ordered):
+def test_ordered(ordered: Ordered):
     """Test post generation are ordered by creation counter."""
     assert ordered.value == "aaa"